+++ app/config/doolox.php
@@ -1,0 +1,17 @@
+<?php
+
+return array(
+
+	/*
+	|--------------------------------------------------------------------------
+	| Allow User Management To Regular Users
+	|--------------------------------------------------------------------------
+	|
+	| When you set this to true, all users can manage their sub-users. When set
+    | to false, only superusers can manage Doolox users.
+	|
+	*/
+
+	'allow_user_management' => true,
+
+);