+++ app/database/migrations/2014_01_18_061536_install.php
@@ -14,6 +14,7 @@ class Install extends Migration {
 
         Schema::create('users', function($table) {
             $table->increments('id');
+            $table->integer('parent_id')->unsigned()->nullable();
             $table->string('email')->unique();
             $table->string('password');
             $table->boolean('superuser');