+++ app/views/dashboard.blade.php
@@ -13,6 +13,12 @@
             {{ Session::get('success') }}
         </div>
 @endif
+@if(Session::has('error'))
+        <div class="alert alert-danger alert-dismissable">
+            <button type="button" class="close" data-dismiss="alert" aria-hidden="true">×</button>
+            {{ Session::get('error') }}
+        </div>
+@endif
     </div>
 </div><!-- /.row -->